requirements.txt
beanstalkc3==0.4.0
requests==2.22.0
